fix: move docer command
