+++ cmd/kubectl-testkube/commands/docker/init.go
@@ -1,7 +1,8 @@
-package pro
+package docker
 
 import (
 	"errors"
+	"fmt"
 	"os"
 
 	"github.com/pterm/pterm"
@@ -9,10 +10,11 @@ import (
 
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common"
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/config"
+	"github.com/kubeshop/testkube/pkg/telemetry"
 	"github.com/kubeshop/testkube/pkg/ui"
 )
 
-func NewDockerCmd() *cobra.Command {
+func NewInitCmd() *cobra.Command {
 	var noLogin bool // ignore ask for login
 	var dockerContainerName, dockerImage string
 	var options common.HelmOptions
@@ -25,9 +27,9 @@ func NewDockerCmd() *cobra.Command {
 	}
 
 	cmd := &cobra.Command{
-		Use:     "docker",
+		Use:     "init",
 		Short:   "Run Testkube Docker Agent and connect to Testkube Pro environment",
-		Aliases: []string{"da", "docker-agent"},
+		Aliases: []string{"install", "agent"},
 		Run: func(cmd *cobra.Command, args []string) {
 			ui.Info("WELCOME TO")
 			ui.Logo()
@@ -136,4 +138,29 @@ func NewDockerCmd() *cobra.Command {
 	cmd.Flags().StringVar(&dockerImage, "docker-image", "kubeshop/testkube-agent"+latestVersion, "Docker image for Testkube Docker Agent")
 
 	return cmd
+}
+
+func sendErrTelemetry(cmd *cobra.Command, clientCfg config.Data, errType string, errorLogs error) {
+	var errorStackTrace string
+	errorStackTrace = fmt.Sprintf("%+v", errorLogs)
+	if clientCfg.TelemetryEnabled {
+		ui.Debug("collecting anonymous telemetry data, you can disable it by calling `kubectl testkube disable telemetry`")
+		out, err := telemetry.SendCmdErrorEvent(cmd, common.Version, errType, errorStackTrace)
+		if ui.Verbose && err != nil {
+			ui.Err(err)
+		}
+
+		ui.Debug("telemetry send event response", out)
+	}
+}
+
+func sendAttemptTelemetry(cmd *cobra.Command, clientCfg config.Data) {
+	if clientCfg.TelemetryEnabled {
+		ui.Debug("collecting anonymous telemetry data, you can disable it by calling `kubectl testkube disable telemetry`")
+		out, err := telemetry.SendCmdAttemptEvent(cmd, common.Version)
+		if ui.Verbose && err != nil {
+			ui.Err(err)
+		}
+		ui.Debug("telemetry send event response", out)
+	}
 }
