+++ cmd/kubectl-testkube/commands/pro.go
@@ -19,7 +19,6 @@ func NewProCmd() *cobra.Command {
 	cmd.AddCommand(pro.NewDisconnectCmd())
 	cmd.AddCommand(pro.NewInitCmd())
 	cmd.AddCommand(pro.NewLoginCmd())
-	cmd.AddCommand(pro.NewDockerCmd())
 
 	return cmd
 }
